refactor: moved createMenu func in helpers
