+++ src/index.js
@@ -1,5 +1,5 @@
 import process from 'node:process';
-import { movesObj } from './helpers/helper.js';
+import { movesObj, createMenu } from './helpers/helper.js';
 //import crypto from 'node:crypto';
 //console.log(crypto.randomBytes(32).toString('hex'))
 
@@ -19,15 +19,6 @@ if (argv.length < 3){
     moves = argv.slice(2);
     console.log(moves);
     console.log(movesObj(moves));
-    function createMenu(arr){
-        let menu = 'Available moves:\n';
-        const constMoves = '0 - exit\n? - help';
-        for (let i=0; i<arr.length; i++){
-            menu += `${i+1} - ${arr[i]}\n`;
-        }
-        menu += constMoves;
-        return menu;
-    }
     console.log(createMenu(moves))
 }
